+++ websocket/model/web_socket_msg.go
@@ -3,7 +3,7 @@ package model
 // Setting 设置表
 type WebSocketMsg struct {
 	ID           uint   `gorm:"default:0;comment:消息ID;NOT NULL" json:"id"`
-	CompanyUuid  uint   `gorm:"default:0;comment:消息ID;NOT NULL" json:"company_uuid"`
+	CompanyUuid  uint64 `gorm:"default:0;comment:消息ID;NOT NULL" json:"company_uuid"`
 	Uid          string `gorm:"default:'';comment:Uuid;NOT NULL" json:"uid"`
 	Msg          string `gorm:"default:'';comment:消息内容;NOT NULL" json:"msg"`
 	Type         string `gorm:"default:'';comment:消息类型;NOT NULL" json:"type"`
